requirements.txt
asn1crypto==0.24.0
backports.shutil-get-terminal-size==1.0.0
cryptography==2.1.4
decorator==4.3.0
enum34==1.1.6
idna==2.6
ipaddress==1.0.17
ipython==5.7.0
ipython-genutils==0.2.0
keyring==10.6.0
keyrings.alt==3.0
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycrypto==2.6.1
Pygments==2.2.0
pygobject==3.26.1
pyxdg==0.25
scandir==1.7
SecretStorage==2.3.1
selenium==3.13.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7

